+++ datadog_checks_base/CHANGELOG.md
@@ -3,6 +3,7 @@
 1.2.2 / Unreleased
 ==================
 
+* [FEATURE] The generic Prometheus check will now send counter as monotonic counter.
 * [BUG] Prometheus requests can use an insecure option
 * [BUG] Correctly handle missing counters/strings in PDH checks when possible
 * [BUG] Fix Prometheus Scrapper logger